added empty shells of all the subsystem methods i need and updated hardware map to include the stopper servo
