+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/gamma/gammaHardwareMapConfig.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.gamma;
 
 import com.qualcomm.hardware.gobilda.GoBildaPinpointDriver;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
@@ -15,34 +15,33 @@ import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
  * All subsystems receive this object in their constructor
  * and pull references from it — they never call hw.get() themselves.
  */
-public class HardwareMapConfig {
+public class gammaHardwareMapConfig {
 
-    // drivetrain (Mecanum, 4 motors)
+    // --- Drivetrain --- (mecanum, 4 motors)
     public DcMotorEx wheel_0;  // front-right (NE)
     public DcMotorEx wheel_1;  // back-right  (SE)
     public DcMotorEx wheel_2;  // back-left   (SW)
     public DcMotorEx wheel_3;  // front-left  (NW)
 
-    // shooter (dual flywheel + feeder servo + stopper servo)
+    // --- Shooter --- (dual flywheel + feeder servo)
     public DcMotorEx shooter_motor_0;
     public DcMotorEx shooter_motor_1;
-    public Servo feeder_servo;
-    public Servo stopper_servo;
+    public Servo     feeder_servo;
 
-    // intake (single roller motor)
+    // --- Intake --- (single roller motor)
     public DcMotorEx intake_motor;
 
-    // turret (rotation motor)
+    // --- Turret --- (rotation motor)
     public DcMotorEx turret_motor;
 
-    // webcam
+    // --- Webcam ---
     public WebcamName webcam;
 
-    // odometry pods management
+    // --- Odometry ---
     public GoBildaPinpointDriver pinpoint;
 
 
-    public HardwareMapConfig(HardwareMap hw) {
+    public gammaHardwareMapConfig(HardwareMap hw) {
 
         // Drivetrain
         wheel_0 = hw.get(DcMotorEx.class, "wheel_0");
@@ -53,8 +52,7 @@ public class HardwareMapConfig {
         // Shooter
         shooter_motor_0 = hw.get(DcMotorEx.class, "shooter_motor_0");
         shooter_motor_1 = hw.get(DcMotorEx.class, "shooter_motor_1");
-        feeder_servo = hw.get(Servo.class, "feeder_servo");
-        stopper_servo = hw.get(Servo.class, "stopper_servo");
+        feeder_servo    = hw.get(Servo.class, "feeder_servo");
 
         // Intake
         intake_motor = hw.get(DcMotorEx.class, "intake_motor");
@@ -62,18 +60,20 @@ public class HardwareMapConfig {
         // Turret
         turret_motor = hw.get(DcMotorEx.class, "turret_motor");
 
-        // Webcam
+        //Webcam
         webcam = hw.get(WebcamName.class, "webcam");
 
-        // Pinpoint odometry
+        //Pinpoint
         pinpoint = hw.get(GoBildaPinpointDriver.class, "pinpoint");
 
-        // Pinpoint calibration
-        pinpoint.setOffsets(-45, -126, DistanceUnit.MM); // pod mounting positions
-        //pinpoint.setTrack(90); // distance between the dead wheels (to calibrate rotation)
+        /* PINPOINT CONFIGURATION
+        These must be set here so the odometry subsystem file starts with a calibrated unit.
+         */
+        pinpoint.setOffsets(-45, -126, DistanceUnit.MM);   // tune these to your actual pod mounting positions
+        //pinpoint.setTrack(90); // distance between the dead wheels (to calibrate rotation
         pinpoint.setEncoderResolution(GoBildaPinpointDriver.GoBildaOdometryPods.goBILDA_4_BAR_POD);  // 2000 CPR, 32mm wheel
         pinpoint.setEncoderDirections(GoBildaPinpointDriver.EncoderDirection.FORWARD,
                 GoBildaPinpointDriver.EncoderDirection.FORWARD);
-        pinpoint.resetPosAndIMU(); // resets to (0,0,0)
+        pinpoint.resetPosAndIMU();
     }
 }
